Enhance ZebinWriter and ByteSink functionality for improved encoding
- Updated the `ZebinWriter` to accept a generic `ByteSink`, allowing for more flexible buffer management during encoding.
- Introduced a new `LimitedSink` implementation for testing, enabling controlled write limits and alignment handling.
- Refactored various tests to utilize the new `writer` function and `SliceEncoder`, ensuring consistency and clarity in encoding operations.
- Improved the `ByteSink` trait with additional implementations for better usability and integration with existing structures.

+++ zebin-macros/src/serialize.rs
@@ -40,7 +40,7 @@ fn field_measure_expr(
                 zebin::utils::macro_helpers::measure_packed_bool_len((#value).len())
             },
             crate::shared::packed::PackedElementKind::U8 => {
-                let bits_lit = bits as u8;
+                let bits_lit = bits;
                 quote! {
                     zebin::utils::macro_helpers::measure_packed_u8_len((#value).len(), #bits_lit)?
                 }
